+++ packages/provider/src/enums/chains.ts
@@ -4,7 +4,12 @@ export enum ChainTypes {
   Near = 3,
 }
 export enum SolanaChains {
-  devnet = 'devnet',
-  testnet = 'testnet',
-  mainnet = 'mainnet',
+  DevNet = 'devnet',
+  TestNet = 'testnet',
+  MainNet = 'mainnet',
+}
+
+export enum NearChains {
+  TestNet = 'testnet',
+  MainNet = 'mainnet',
 }