+++ packages/provider/src/providers/base-solana.ts
@@ -1,6 +1,4 @@
 import { PublicKey } from '@solana/web3.js'
-import Web3 from 'web3'
-import { HttpProvider } from 'web3-core'
 
 import { ChainTypes, ProviderEvents, SolanaChains } from '@/enums'
 import {
@@ -33,14 +31,12 @@ export class BaseSolanaProvider
   implements ProviderProxy
 {
   readonly #provider: SolanaProvider
-  #web3: Web3
   #chainId?: ChainId
   #address?: string
 
   constructor(provider: RawProvider) {
     super()
-    this.#web3 = new window.Web3(provider as unknown as HttpProvider)
-    this.#provider = (<unknown>this.#web3?.currentProvider) as SolanaProvider
+    this.#provider = provider as SolanaProvider
   }
 
   get chainType(): ChainTypes {
@@ -66,7 +62,7 @@ export class BaseSolanaProvider
   async init(): Promise<void> {
     this.#setListeners()
     this.#address = getAddress(this.#provider.publicKey)
-    this.#chainId = SolanaChains.devnet
+    this.#chainId = SolanaChains.DevNet
 
     this.emitInitiated({
       chainId: this.#chainId,