+++ packages/provider/src/providers/index.ts
@@ -1,4 +1,6 @@
 export * from './coinbase'
 export * from './metamask'
+export * from './near'
 export * from './phantom'
+export * from './solflare'
 export * from './solflare'