+++ packages/provider/src/types/index.ts
@@ -1,5 +1,7 @@
 export * from './chain'
+export * from './chain'
 export * from './ethereum'
+export * from './near'
 export * from './provider'
 export * from './provider-event-bus'
 export * from './solana'